+++ Functions/GeneralFunction.cs
@@ -7,25 +7,25 @@ using System.Threading.Tasks;
 namespace NeuralNetwork.Functions {
 	public abstract class GeneralFunction {
 
-		private static bool betw(float v, float v1, float v2) {
+		private static bool betw(double v, double v1, double v2) {
 			return v1 > v2 ? v2 <= v && v <= v1 : v1 <= v && v <= v2;
 		}
 
-		public float this[float x] {
+		public double this[double x] {
 			get {
 				return Function(x);
 			}
 		}
 
-		protected float param;
+		protected double param;
 
-		public abstract float Function(float x);
+		public abstract double Function(double x);
 
-		public virtual float Inverse(float y) {
-			float x = -10f;
-			float step = 1f;
-			float oldVal = Function(x);
-			float newVal = Function(x);
+		public virtual double Inverse(double y) {
+			double x = -10f;
+			double step = 1f;
+			double oldVal = Function(x);
+			double newVal = Function(x);
 			while(step >= 0.001f) {
 				while(!betw(y, oldVal, newVal)) {
 					x += step;
@@ -40,12 +40,12 @@ namespace NeuralNetwork.Functions {
 			return x + ( step / 2f );
 		}
 
-		public virtual bool SetParameterFor(float x, float y) {
-			float oldP = param;
+		public virtual bool SetParameterFor(double x, double y) {
+			double oldP = param;
 			param = -10f;
-			float step = 1f;
-			float oldVal = Function(x);
-			float newVal = Function(x);
+			double step = 1f;
+			double oldVal = Function(x);
+			double newVal = Function(x);
 			while(step >= 0.001f) {
 				step /= 10f;
 				while(!betw(y, oldVal, newVal)) {
